+++ src/pages/adminuseredit/historial/index.js
@@ -6,7 +6,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faDownload } from '@fortawesome/free-solid-svg-icons';
 import { useFirebaseApp } from 'reactfire';
 import { StyledPanel, StyleHeader } from './styled';
-import { CSVLink } from 'react-csv';
+import ExportReactCSV from '../../dowloadData/index';
 
 const StyledBadge = styled(Badge)`
     color: #09d3ac;
@@ -59,32 +59,6 @@ export default function HistoryUser({ user }) {
     const db = firebase.firestore();
     const [history, setHistory] = useState([]);
     const [tableData, setTableData] = useState();
-    const headers = [
-        { label: 'ID', key: 'id' },
-        { label: 'Nombre', key: 'name' },
-        { label: 'Fecha', key: 'package.creation_date' },
-        { label: 'Guia', key: 'rastreo' },
-        { label: 'Nombre Origen', key: 'sender_addresses.name' },
-        { label: 'Telefono Origen', key: 'sender_addresses.phone' },
-        { label: 'Calle Origen', key: 'sender_addresses.street_number' },
-        { label: 'Colonia Origen', key: 'sender_addresses.neighborhood' },
-        { label: 'Código Postal Origen', key: 'sender_addresses.codigo_postal' },
-        { label: 'Referencias Origen', key: 'sender_addresses.place_reference' },
-        { label: 'Nombre Destino', key: 'receiver_addresses.name' },
-        { label: 'Telefono Destino', key: 'receiver_addresses.phone' },
-        { label: 'Calle Destino', key: 'receiver_addresses.street_number' },
-        { label: 'Colonia Destino', key: 'receiver_addresses.neighborhood' },
-        { label: 'Código Postal Destino', key: 'receiver_addresses.codigo_postal' },
-        { label: 'Referencias Destino', key: 'receiver_addresses.place_reference' },
-        { label: 'Paquete', key: 'package.content_description' },
-        { label: 'Paquete Peso', key: 'package.weight' },
-        { label: 'Paquete Largo', key: 'package.height' },
-        { label: 'Paquete Ancho', key: 'package.width' },
-        { label: 'Paquete Alto', key: 'package.depth' },
-        { label: 'Servicios', key: 'supplierData.Supplier' },
-        { label: 'Costo', key: 'supplierData.Supplier_cost' },
-    ];
-
     useEffect(() => {
         if (user) {
             let dataGuias = [];
@@ -140,16 +114,7 @@ export default function HistoryUser({ user }) {
             <StyleHeader>
                 <Row className="row-header">
                     <h2>Historial de envíos</h2>
-                    <Button variant="destructive" className="rainbow-m-around_medium">
-                        <CSVLink
-                            data={history}
-                            headers={headers}
-                            filename={'historial-de-envios.csv'}
-                            target="_blank"
-                        >
-                            Descargar archivo
-                        </CSVLink>
-                    </Button>
+                    <ExportReactCSV data={history} />
                 </Row>
             </StyleHeader>
             <div className="rainbow-p-bottom_large rainbow-p-top_large">
