+++ src/pages/adminusers/styled.js
@@ -22,5 +22,14 @@ export const StyledAusers = styled.div`
         justify-content: space-between;
         align-items: center;
         margin: 0;
+
+        h1 {
+            margin-bottom : 0;
+        }
+        button {
+            background-color: #ab0000;
+            a {
+                color: white;
+            }
     }
 `;
