+++ src/pages/dowloadData/index.js
@@ -1,0 +1,44 @@
+import React from 'react';
+import { CSVLink } from 'react-csv';
+import { Button } from 'react-rainbow-components';
+
+export default function ExportReactCSV({ data }) {
+    console.log('data csv', data);
+    const headers = [
+        { label: 'ID', key: 'id' },
+        { label: 'Nombre', key: 'name' },
+        { label: 'Fecha', key: 'package.creation_date' },
+        { label: 'Guia', key: 'rastreo' },
+        { label: 'Nombre Origen', key: 'sender_addresses.name' },
+        { label: 'Telefono Origen', key: 'sender_addresses.phone' },
+        { label: 'Calle Origen', key: 'sender_addresses.street_number' },
+        { label: 'Colonia Origen', key: 'sender_addresses.neighborhood' },
+        { label: 'Código Postal Origen', key: 'sender_addresses.codigo_postal' },
+        { label: 'Referencias Origen', key: 'sender_addresses.place_reference' },
+        { label: 'Nombre Destino', key: 'receiver_addresses.name' },
+        { label: 'Telefono Destino', key: 'receiver_addresses.phone' },
+        { label: 'Calle Destino', key: 'receiver_addresses.street_number' },
+        { label: 'Colonia Destino', key: 'receiver_addresses.neighborhood' },
+        { label: 'Código Postal Destino', key: 'receiver_addresses.codigo_postal' },
+        { label: 'Referencias Destino', key: 'receiver_addresses.place_reference' },
+        { label: 'Paquete', key: 'package.content_description' },
+        { label: 'Paquete Peso', key: 'package.weight' },
+        { label: 'Paquete Largo', key: 'package.height' },
+        { label: 'Paquete Ancho', key: 'package.width' },
+        { label: 'Paquete Alto', key: 'package.depth' },
+        { label: 'Servicios', key: 'supplierData.Supplier' },
+        { label: 'Costo', key: 'supplierData.Supplier_cost' },
+    ];
+    return (
+        <Button variant="destructive" className="rainbow-m-around_medium">
+            <CSVLink
+                data={data}
+                headers={headers}
+                filename={'historial-de-envios.csv'}
+                target="_blank"
+            >
+                Descargar archivo
+            </CSVLink>
+        </Button>
+    );
+}
